+++ src/widgets/member/index.ts
@@ -1,0 +1,1 @@
+export { SettingList } from './ui/SettingList/SettingList';
