+++ src/widgets/member/ui/SettingList/ThicknessSetMenu/ThicknessSetMenu.tsx
